리팩토링
1.주석 삭제

2. RunningGuide 객체에 tag 추가

3. LocationManager, RunningGuideItem 폴더 분류 하여 이동
