+++ GatheRunner/Presentaion/Running/LocationManager/LocationManager.swift
@@ -1,0 +1,25 @@
+import Foundation
+import MapKit
+import SwiftUI
+
+// MARK: 에러 예외처리 필요 -> 구현 예정
+class LocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
+    @Published var region = MKCoordinateRegion()
+    private let manager = CLLocationManager()
+    
+    override init() {
+        super.init()
+        manager.delegate = self
+        manager.desiredAccuracy = kCLLocationAccuracyBest
+        manager.requestWhenInUseAuthorization()
+        manager.startUpdatingLocation()
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        locations.last.map {
+            let center = CLLocationCoordinate2D(latitude: $0.coordinate.latitude, longitude: $0.coordinate.longitude)
+            let span = MKCoordinateSpan(latitudeDelta: 0.004, longitudeDelta: 0.004)
+            region = MKCoordinateRegion(center: center, span: span)
+        }
+    }
+}
